EAG_V2/Assignment_5/Math_Agent/pyproject.toml
[project]
name = "math-agent"
version = "0.1.0"
description = "Mathematical problem solver using Gemini AI and MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "sympy>=1.12",
    "requests>=2.31.0",
]
